-Added Endless Pouch of Life -Added Titanic Gatli Stynger -Fixed type in Accessories directory -Improved Shoot method in ModPlayer
