+++ Decimation.csproj
@@ -69,30 +69,31 @@
     <Compile Include="Buffs\Warlock.cs" />
     <Compile Include="DecimationPlayer.cs" />
     <Compile Include="DecimationWorld.cs" />
-    <Compile Include="Items\Accesories\CrystalSkull.cs" />
-    <Compile Include="Items\Accesories\DeadeyesQuiver.cs" />
-    <Compile Include="Items\Accesories\EnchantedFocuser.cs" />
-    <Compile Include="Items\Accesories\EnergyFocuser.cs" />
-    <Compile Include="Items\Accesories\FireAmulet.cs" />
-    <Compile Include="Items\Accesories\Focuser.cs" />
-    <Compile Include="Items\Accesories\JestersQuiver.cs" />
-    <Compile Include="Items\Accesories\NecrosisStone.cs" />
-    <Compile Include="Items\Accesories\RangersPouch.cs" />
-    <Compile Include="Items\Accesories\RangersQuiver.cs" />
-    <Compile Include="Items\Accesories\RedHotShackle.cs" />
-    <Compile Include="Items\Accesories\SlimeAmulet.cs" />
-    <Compile Include="Items\Accesories\SlimeBracelet.cs" />
+    <Compile Include="Items\Accessories\CrystalSkull.cs" />
+    <Compile Include="Items\Accessories\DeadeyesQuiver.cs" />
+    <Compile Include="Items\Accessories\EnchantedFocuser.cs" />
+    <Compile Include="Items\Accessories\EnergyFocuser.cs" />
+    <Compile Include="Items\Accessories\FireAmulet.cs" />
+    <Compile Include="Items\Accessories\Focuser.cs" />
+    <Compile Include="Items\Accessories\JestersQuiver.cs" />
+    <Compile Include="Items\Accessories\NecrosisStone.cs" />
+    <Compile Include="Items\Accessories\RangersPouch.cs" />
+    <Compile Include="Items\Accessories\RangersQuiver.cs" />
+    <Compile Include="Items\Accessories\RedHotShackle.cs" />
+    <Compile Include="Items\Accessories\SlimeAmulet.cs" />
+    <Compile Include="Items\Accessories\SlimeBracelet.cs" />
     <Compile Include="Items\Antidote.cs" />
     <Compile Include="Decimation.cs" />
     <Compile Include="Items\BloodiedEssence.cs" />
     <Compile Include="Items\Boss\Arachnus\ArachnusBag.cs" />
     <Compile Include="Items\CommanderPotion.cs" />
-    <Compile Include="Items\DenziumIngot.cs" />
+    <Compile Include="Items\DenziumBar.cs" />
     <Compile Include="Items\DesertDessert.cs" />
     <Compile Include="Items\DuneWormMask.cs" />
     <Compile Include="Items\DuneWormTreasureBag.cs" />
     <Compile Include="Items\DuneWormTrophey.cs" />
     <Compile Include="Items\EnchantedMushroom.cs" />
+    <Compile Include="Items\Accessories\EndlessPouchofLife.cs" />
     <Compile Include="Items\Placeable\DenziumOre.cs" />
     <Compile Include="Items\Placeable\DenziumWall.cs" />
     <Compile Include="Items\Tools\GildedSickle.cs" />
@@ -127,6 +128,7 @@
     <Compile Include="Items\Weapons\Sling.cs" />
     <Compile Include="Items\Weapons\SlingshotWood.cs" />
     <Compile Include="Items\Weapons\TheGreatwoodSword.cs" />
+    <Compile Include="Items\Weapons\TitanicGatliStynger.cs" />
     <Compile Include="Items\Weapons\TitanicLongsword.cs" />
     <Compile Include="Items\Weapons\TitanicPike.cs" />
     <Compile Include="Items\Weapons\TitanicRepeater.cs" />
@@ -145,6 +147,7 @@
     <Compile Include="Projectiles\TitanicPikeProjectile.cs" />
     <Compile Include="Projectiles\TitanicStyngerBolt.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="References.cs" />
     <Compile Include="Structures\ShrineoftheMoltenOne.cs" />
     <Compile Include="Tiles\ChlorophyteAnvil.cs" />
     <Compile Include="Tiles\DenziumOre.cs" />
