--- Items/Accesories/DELETE
@@ -1,1 +1,0 @@
-there are better ways of doing this but there are some files I don't want uploaded just yet
