+++ Items/Accessories/JestersQuiver.png
