+++ Items/Accessories/SlimeAmulet.png
