+++ Items/Accessories/SlimeBracelet.png
